Checkpoint: the board is running!
Rebuilt the bootloader and monitor with updated addressing and
linker script. Running out the SRAM card in slot 0 for now. UART and
IDE seem to be working. No interrupts, timers, PS/2 or any of the
FPGA peripherals are available. Except the LED, which works. :)

Next up will probably be to port/rewrite the old DRAM controller.

Verilator in lint mode is used as a sanity check before passing it to
Quartus. Will look at some testbenches at some point. Trying to keep
MAXI030Core.v as a place to instantiate modules with very little
logic.