Added basic interrupt routing and a timer
The timer has two registers, a starting count (32 bit) and a control
register (8 bit). Currently the control register is erroniously
readable. The interrupts are basic autovectoring, as they were before,
and default to being masked/disabled.

For consistency, all registers are passed in as 32 bit to the subsytem
(LED, timer, whatever). This is done because some subsystems, like
the timer, will have a mix of registers defined in a single module.